requirements.txt
fastapi
uvicorn
psycopg
pandas
python-dotenv
my-sql-connector
requests
streamlit